Create rtsp-baz-url-brute.nse
This is a tweaked version of the default `rtsp-url-brute.nse` script that comes with nmap.
I've named it as `rtsp-baz-url-brute.nse`, because that was the only way I could get nmap to use my version, instead of the default one.

This version differs from the `rtsp-url-brute.nse`, in that:
1. It scans on ports: 554, 8554 and 10554
2. It displays the port number in the results
And

No warranty given, or implied. Don't do crime. Friends don't let friends run RTSP cams on networks with UPnP 'on' by default.